Improve game finished logic
Intoduced PlayerState: AwaitingMove, ConfirmedChange, Skipped.

When a player confirms change story points, all other players state
changes to AwaitingMove. When there are no other players with
AwaitingMove state, the game is finished.

'phase' is now sent in the MOVE_CONFIRMED event (aswell as
in the PLAYER_SKIPPED event as before).